Improve some scripts

+++ bin/make_oanda_env_file.sh
@@ -6,11 +6,17 @@ cd "$dirpath"/..
 
 ENV_FILE_PATH="env/oanda_secret.env"
 
-API_TOKEN=$(bin/get_oanda_api_token.sh) || exit 1 
+API_TOKEN=$(bin/get_oanda_api_token.sh) || exit 1
 ACCOUNT_ID=$(bin/get_oanda_account_id.sh) || exit 1
 
-[ -z API_TOKEN ] || echo "Can't retrieve API_TOKEN"; exit 3
-[ -z ACCOUNT_ID ] || echo "Can't retrieve ACCOUNT_ID"; exit 3
+if [ -z $API_TOKEN ]; then 
+    echo "Can't retrieve API_TOKEN"
+    exit 1
+fi
+if [ -z $ACCOUNT_ID ]; then 
+    echo "Can't retrieve ACCOUNT_ID"
+    exit 1
+fi
 
 rm -f $ENV_FILE_PATH || exit 2
 touch $ENV_FILE_PATH || exit 2
